fix: Fix ctrl-tab falling out of sync after dragging a tab
Dragging a sortable seems to not update the list it's based on, only its
internal list (and thus only the rendering order, which breaks the
ctrl+tab hotkeys). There's a Vue version of Sortable.js (Vue.Draggable)
that seems to fix this but this is a not-terrible workaround for now.