feature: Adjustable Automatic reply (#85)
* Implement custom message option for auto-reply

Functional, implements a custom auto-reply message optionally and fixes the link for the original message in the process.

Needs a little bit more text to nicely instruct some usage.

* Formatting fixes

Replaces a few errant tabbings of my doing with spaces.

* Made UI of this feature a bit better, turned off debug code, made a default-setting in the box based on original message

Moves the auto-reply section down to the bottom and gives it some proper writing to indicate how to use it and that BBCode works as well as the fact that people can bypass the message.

Re-implemented the proper check that I was bypassing to test this feature's functionality (the one I commented out with a big warning to myself and almost forgot).

Also went ahead and made a default setting for the box just so that there's some extra info to work with if it's the first time someone uses it.

* Prettier ran

Gave prettier a quick run, cleaned things up.

* Dot beside names with original gender colour when using a custom colour (optional)

Puts a dot beside people with custom colours for their names so you can still easily spot the gender while preserving their custom name.

Appears in chat, sidebar, and at the top of private conversations.

Has included option to enable/disable it.

Had to expand the sidebar by 10 pixels to accomodate a name with it that was very long.

* Prettier run

Discarded the changes to UserView.vue to preserve BBCode functionality and not mess things up.

* Implements an optional gender icon on everyone.

Largely based on the original colour dot code, pretty simple change. Doesn't do the colour thing, adds an icon for the gender to everyone in chat, on the right sidebar, and in searches.

Aimed at a little combo of colourblind assistance and achieving the same goal as the gender dot originally did.

* Undoing _chat.scss change

This change only really mattered for single line names that were incredibly long so just undoing it since it's an edge case. Most people separate with spaces and underscores anyway and the client handles that okay.

* Made offline people return a times symbol (x) instead of a gendered symbol

Realised it marked everyone currently offline with the same icon as anyone without a gender, gave that a quick change so it is now more obvious that they haven't just removed their gender and instead are just offline.

It makes this change as people go online/offline so added bonus is that you can see when people are offline in rooms much quicker since name colours do not change by default.

Maybe should implement that as a change too, make offline people immediately go grey rather than keep their coloured name.

* Swapped both purple gender icons around

Small change, just that.

* Makes marker appear/disappear in chat like avatar

Decently simple, does what it says on the tin. Console messages require a channel change. Still needs a channel change and a relog for the sidebar to properly respect it.

* Prettier run

Just doing this quick before I forget to

* Makes span appear and disappear properly everywhere, also adds an optional colour change for offline users. Needs more testing

Implemented a get into UserView that grabs the specific wanted setting.

Removed some references to a prop that no longer exists as a consequence of doing things this way.

Also added an option to change colours on anyone who logs off to make them consistent with users who are offline in logs. Primarily added for a little consistency with the offline icon that comes with the gender markers.

* Prettier run

* Updated with a slightly more performance-friendly solution

Does the same as before, changes it almost everywhere instantly.

This route makes it not update in BBCode instantly but a channel change will fix that, it also doesn't update in searches until you do a new search with the setting turned on.

It is also no longer visible in the user's ads tab or shared channels tab. There were a few issues with spacing there and you can still see their gender icon on the main name from there.

* Removed a commented portion

* Prettier Run

* Small fix for RecentConversations and CharacterPreview

RecentConversations did not have the new prop integrated, it does now.

Also put the genderClass into CharacterPreview so you see the marker when hovering over someone. This one isn't wholly necessary considering that the gender is written there but the name is coloured there so I figured it wise to put the marker there too.

* Prettier Run

* Update UserView.vue

* Fixes auto-reply settings being on the same line

* Undid changes to CharacterPreview

Info is already available there and I honestly just didn't think it looked that great when it was put there.

* Prettier Run

* Revert "Merge pull request #9 from BootsieWootsie/feature/gender-markers"

This reverts commit 8c159be1e93d8edd172cdb8e56e5ba5de4d8a36a, reversing
changes made to 4e07851e3dbb0ad30ea0503c26a30f508e56a6fa.

* Options for this now grey out when autoreply is off, also utilises the editor from AdCenter as suggested.

As mentioned, this pulls in the editor from bbcode, utilised in AdCenter. FatCat's suggestion here was far cleverer than my implementation so thank you for that! Also gave me a good place to look to see where to yoink that from with an example of implementation.

It is limited to a fixed value of 10k (5x smaller than the actual message limit in private) on the editor currently rather than importing the maxLength, technically not update resistant but I doubt the max length is changing anytime soon nor that anyone will utilise the full 10k for this ever.

The limit of 10k also leaves it in a position that if those are limited ever, it likely will not dip that far below 50k for private messages which is the only place this should ever fire.

This also makes the custom settings grey out when autoreply is disabled.  This applies both to the requested section and the editor too.

Text describing how this works a little has been updated to reflect the inclusion of an editor and the preview, as well as to include the character limit.

* Prettier run

* Make editor disable when either autoReply or autoReplyCustom is disabled. Also renames the new vars for a bit of consistency and length regulation.

---------

Co-authored-by: FatCatClient <82377182+FatCatClient@users.noreply.github.com>
Co-authored-by: Wired Rose <rose@rosely.me>